moved images to assets/images
