+++ src/components/Header/DocumentTextbox.tsx
@@ -6,7 +6,7 @@ import {
 } from 'react';
 import { useAppDispatch, useAppSelector } from '@/store/hooks';
 
-import IconDocument from '@/assets/icon-document.svg?react';
+import IconDocument from '@/assets/images/icon-document.svg?react';
 import {
     updateDocumentName,
     selectActiveDocumentMetadata,
