+++ src/components/Header/SidebarSwitch.tsx
@@ -1,5 +1,5 @@
-import IconMenu from '@/assets/icon-menu.svg?react';
-import IconClose from '@/assets/icon-close.svg?react';
+import IconMenu from '@/assets/images/icon-menu.svg?react';
+import IconClose from '@/assets/images/icon-close.svg?react';
 
 type Props = {
     expanded: boolean;
